Handle double create of Env
We create and maintain single instances of all Env objects. Snakemake
treats them as normal objects, and creates additional instances for
environments to be installed once the DAG is complete. So to avoid
triggering an error indicating double definition, we need to
essentially implement an multi-singleton solution for the Env objects.