Fix endless loop
Getting a config value by index with offset > len() did not lead to an
IndexError. The new sequence default code for self.index(val) runs
into an endless loop if this is the case, leading to endless loop
when trying to so something simple like `self.pairnames.index('R2')`
and R2 is not present.